+++ js/utils.js
@@ -1,4 +1,26 @@
 /**
+ * Отправляет HTTP-запрос на указанный url с настройками options
+ * @param {string} url
+ * @param {RequestInit} [options]
+ * @returns
+ */
+async function request(url, options) {
+  const response = await fetch(url, options);
+  if (!response.ok) {
+    throw new Error(`Ошибка получения данных!\n${response.status}: ${response.statusText}`);
+  }
+  return response.json();
+}
+
+const debounce = (callback, timeoutDelay) => {
+  let timeoutId;
+  return (...rest) => {
+    clearTimeout(timeoutId);
+    timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
+  };
+};
+
+/**
  * Генерирует случайное число из диапазона от min до max.
  * @param {number} min
  * @param {number} max
@@ -36,19 +58,11 @@ const getUnicRandomNumber = (min, max) => {
 };
 
 /**
- * Возвращает массив из count уникальных случайных чисел из диапазона от min до max
- * @param {number} min
- * @param {number} max
- * @param {number} count
- * @return {Array<number>}
+ * Разбивает строку слов на массив строк
+ * @param {string} string
+ * @returns {Array<string>}
  */
-const getSomeRandomNumbers = (min, max, count) => {
-  const someRandomNumbers = [];
-  const getUnicNumber = getUnicRandomNumber(min, max);
-  for (let i = 0; i < count; i++) {
-    someRandomNumbers.push(getUnicNumber());
-  }
-  return someRandomNumbers;
-};
+const segmentWords = (string) => string.toLowerCase().split(' ').filter(Boolean);
 
-export {getRandomInteger, getUnicRandomNumber, getSomeRandomNumbers};
+
+export {request, debounce, getRandomInteger, getUnicRandomNumber, segmentWords};
